Deploy 24/5

+++ app/Http/Controllers/CarreraController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Exception;
 use Illuminate\Http\Request;
 use App\Models\Carrera;
 use App\Models\Grado;
@@ -13,9 +14,20 @@ use LdapRecord\Query\Events\Read;
 
 class CarreraController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
+        if($request->eliminado)
+        {
+            return response()->json(Carrera::onlyTrashed()->get());
+        }
         return response()->json(Carrera::all()->load('grado'));
+    }
+
+    public function activar($id){
+        $carrera = Carrera::onlyTrashed()->find($id);
+        $carrera->restore();
+        Grado::onlyTrashed()->where('carrera_id', $id)->restore();
+        return response()->json($carrera);
     }
 
     public function show($id)
@@ -31,21 +43,20 @@ class CarreraController extends Controller
             'categoria' => 'required|string|max:30',
             'grados' => 'array',
         ]);
-        $carrera=Carrera::create($request->all());
+        $carrera = Carrera::create($request->all());
 
-        if($request->grados)
-        {
+        if ($request->grados) {
             self::agregarCarreraGrados($request->grados, $carrera);
         }
 
         RegistrosController::store("CARRERA", $request->header('token'), "CREATE", $request->nombre);
-        return response()->json($carrera, 201);
+        return response()->json($carrera->load('grado'), 201);
     }
 
     public function update(Request $request, $id)
     {
         $request->validate([
-            'nombre' => 'string|unique:carreras,nombre',
+            'nombre' => 'string|required',
             'plan' => 'string|max:4',
             'categoria' => 'string|max:30',
             'grados' => 'array',
@@ -65,31 +76,58 @@ class CarreraController extends Controller
 
     public function destroy(Request $request, $id)
     {
-        if($request->idGrado)
-        {
-                $grado = Grado::findOrFail($request->idGrado);
-                $grado->delete();
-                return response()->json(['status' => 'success']);
+        if(empty($id)){
+            return response()->json(['status' => "Bad request"], 404);
         }
-
+       try { 
             $carrera = Carrera::findOrFail($id);
-            $carrera->grado()->delete();
+            foreach($carrera->grado as $grado){
+                $grado->grupos()->delete();
+                $grado->materias()->detach();
+                $grado->delete();
+            }
             $carrera->delete();
+
             return $carrera;
-     
-        RegistrosController::store("CARRERA", $request->header('token'), "DELETE", $carrera->nombre);
-        
+            RegistrosController::store("CARRERA", $request->header('token'), "DELETE", $carrera->nombre);
+       } catch (Exception $e) {
+            return response()->json(['status' => "Error al eliminar carrera"], 400);
+        } 
+
+
+    }
+
+    public function destroyGrado($id, $idGrado, Request $request)
+    {
+        $carrera = Carrera::findOrFail($id);
+        $grado = $carrera->grado->find($idGrado);
+
+        if (empty($grado)) {
+            return response()->json(['status' => "Grado no encontrado"], 404);
+        }
+        try {
+            $grado->delete();
+        } catch (Exception $e) {
+            return response()->json(['status' => "Error al eliminar grado"], 403);
+        }
+        RegistrosController::store("CARRERA GRADO", $request->header('token'), "DELETE", $carrera->nombre . " " . $grado->grado);
+        return response()->json(['status' => 'success']);
     }
 
     public function agregarCarreraGrados($grados, $carrera)
     {
         foreach ($grados as $grado) {
-            if($carrera->grado()->where('grado', $grado)->first()){
+            if ($carrera->grado()->where('grado', $grado)->first()) {
                 continue;
             }
-           $carrera->grado()->create([
-            'grado' => $grado
-        ]);
+            $g=Grado::onlyTrashed()->where('carrera_id', $carrera->id)->where('grado', $grado)->first();
+            if ($g){
+                $g->restore();
+                continue;
+            }
+            $carrera->grado()->create([
+                'grado' => $grado
+            ]);
         }
     }
 
@@ -100,10 +138,10 @@ class CarreraController extends Controller
         }
     }
 
-    public function agregarCarreraGrupos (Request $request)
+    public function agregarCarreraGrupos(Request $request)
     {
-        foreach($request->grupo_id as $grupo)
-        grupos_pertenecen_carrera::create($request->carrera_id, $request->grado_id, $grupo);
+        foreach ($request->grupo_id as $grupo)
+            grupos_pertenecen_carrera::create($request->carrera_id, $request->grado_id, $grupo);
     }
-    
+
 }
